+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_moving_single_control_point.dart
@@ -19,7 +19,7 @@ class MPTH2FileEditStateMovingSingleControlPoint extends MPTH2FileEditState
 
     if (MPTH2FileEditStateClearSelectionOnExitMixin.selectionStatesTypes
         .contains(nextStateType)) {
-      if (!MPTH2FileEditStateSingleLineEdit.singleLineEditModes.contains(
+      if (!MPTH2FileEditStateEditSingleLine.singleLineEditModes.contains(
         nextStateType,
       )) {
         selectionController.clearSelectedLineSegments();